rename /controllers/profile.js to /controllers/profileRoutes.js
